API_on_cloud/Python_script_for_stats_API_request/requirements.txt
requests==2.26.0
matplotlib==3.4.3
Pillow==8.3.2
pandas==1.3.3
numpy==1.21.2
